requirements.txt
Django
Pillow
djangorestframework
django-cors-headers
djangorestframework-jwt
djangorestframework-simplejwt